udta/udta-server/requirements.txt
attrs==18.2.0
impyla==0.14.1
mysqlclient==1.3.13
redis==2.10.6
uvloop==0.11.2
